refactor: renamed linters.ts to compilers.ts
